fix #854 Rework sitemap builder so that it only uses the atlas document. Not longer watches for changes to nunaliit_custom.js.
