+++ nunaliit2-couch-metadata/src/test/resources/navigation/levels_3.json
@@ -1,83 +1,98 @@
 {
-  "_id": "navigation.demo",
-  "_rev": "2-d13d52f53d5c257919516b2126b1d65f",
-  "nunaliit_schema": "navigation",
-  "nunaliit_navigation": {
-    "nunaliit_type": "navigation",
-    "title": {
-      "nunaliit_type": "localized",
-      "en": "Atlas",
-      "fr": "L'Atlas"
-    },
-    "items": [
-      {
-        "title": {
-          "nunaliit_type": "localized",
-          "en": "Default Module",
-          "fr": "Module de Base"
-        },
-        "module": "module.test5_map",
-        "items": [
-          {
-            "module": "module.test5_map"
-          }
-        ]
-      },
-      {
-        "title": {
-          "nunaliit_type": "localized",
-          "en": "Canvas Module",
-          "fr": "Module de Canvas"
-        },
-        "module": "module.test5_canvas",
-        "items": [
-          {
-            "module": "module.test5_canvas"
-          },
-          {
-            "title": {
-              "nunaliit_type": "localized",
-              "en": "Default Module",
-              "fr": "Module de Base"
-            },
-            "module": "module.test5_map",
-            "items": [
-              {
-                "module": "module.test5_map"
-              },
-              {
-                "title": {
-                  "nunaliit_type": "localized",
-                  "en": "Tools",
-                  "fr": "Outils"
-                },
-                "href": "./tools/index.html",
-                "items": [
-                  {
-                    "module": "module.test5_map"
-                  },
-                  {
-                    "module": "module.test5_anothermodule"
-                  }
-                ]
-              }
-            ]
-          }
-        ]
-      },
-      {
-        "title": {
-          "nunaliit_type": "localized",
-          "en": "Tools",
-          "fr": "Outils"
-        },
-        "href": "./tools/index.html"
-      }
-    ]
-  },
+  "_id": "atlas",
+  "_rev": "7-cfe4bd758d60e4e10cdc099c363768bd",
   "nunaliit_manifest": {
     "attachments": {},
-    "digest": "/KxzdWOcguFvFh7Z6AdlpqHnKqw=",
+    "digest": "vx1oqqsVrU82GmRPyb+OF2NSEI8=",
     "type": "SHA-1"
+  },
+  "nunaliit_last_updated": {
+    "nunaliit_type": "actionstamp",
+    "name": "admin",
+    "action": "updated",
+    "time": 1580246898155
+  },
+  "nunaliit_schema": "atlas",
+  "nunaliit_atlas": {
+    "nunaliit_type": "atlas",
+    "title": "Atlas Title",
+    "nunaliit_navigation": {
+      "nunaliit_type": "navigation",
+      "title": {
+        "nunaliit_type": "localized",
+        "en": "Atlas",
+        "fr": "L'Atlas"
+      },
+      "items": [
+        {
+          "title": {
+            "nunaliit_type": "localized",
+            "en": "Default Module",
+            "fr": "Module de Base"
+          },
+          "module": "module.test5_map",
+          "items": [
+            {
+              "module": "module.test5_map"
+            }
+          ]
+        },
+        {
+          "title": {
+            "nunaliit_type": "localized",
+            "en": "Canvas Module",
+            "fr": "Module de Canvas"
+          },
+          "module": "module.test5_canvas",
+          "items": [
+            {
+              "module": "module.test5_canvas"
+            },
+            {
+              "title": {
+                "nunaliit_type": "localized",
+                "en": "Default Module",
+                "fr": "Module de Base"
+              },
+              "module": "module.test5_map",
+              "items": [
+                {
+                  "module": "module.test5_map"
+                },
+                {
+                  "title": {
+                    "nunaliit_type": "localized",
+                    "en": "Tools",
+                    "fr": "Outils"
+                  },
+                  "href": "./tools/index.html",
+                  "items": [
+                    {
+                      "module": "module.test5_map"
+                    },
+                    {
+                      "module": "module.test5_anothermodule"
+                    }
+                  ]
+                }
+              ]
+            }
+          ]
+        },
+        {
+          "title": {
+            "nunaliit_type": "localized",
+            "en": "Tools",
+            "fr": "Outils"
+          },
+          "href": "./tools/index.html"
+        }
+      ]
+    },
+    "nunaliit_manifest": {
+      "attachments": {},
+      "digest": "/KxzdWOcguFvFh7Z6AdlpqHnKqw=",
+      "type": "SHA-1"
+    }
   }
 }
